pyproject.toml
[tool.poetry]
name = "aiohttp-demo-backend"
version = "0.1.0"
description = ""
authors = ["Sergey Morichev <s.morichev@yandex.ru>"]
readme = "README.md"
packages = [{include = "aiohttp_demo_backend"}]

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.9"
asyncpg = "^0.27.0"
redis = "^4.5.4"
gunicorn = "^20.1.0"
alembic = "^1.10.3"
aiohttp = "^3.8.4"
pydantic = "^1.10.7"
python-dotenv = "^1.0.0"
aiohttp-security = "^0.4.0"
aiohttp-session = "^2.12.0"

[tool.poetry.group.dev.dependencies]
wemake-python-styleguide = "^0.17.0"
mypy = "^1.2.0"
pytest = "^7.3.1"
coverage = "^7.2.3"
black = "^23.3.0"
pytest-asyncio = "^0.21.0"
freezegun = "^1.2.2"
asyncpg-stubs = "^0.27.0"
pytest-aiohttp = "^1.0.4"

[tool.black]
line-length = 80
target-version = ["py311"]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
